requirements.txt
uvicorn
pydantic
fastapi
pycryptodome
python-multipart
requests
cryptography